% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_FPKM_to_GR.R
\name{add_FPKM_to_GR}
\alias{add_FPKM_to_GR}
\title{Add FPKM values to a genomic ranges structure}
\usage{
add_FPKM_to_GR(GR, FPKM, FPKMgene, columns, FPKMsymbol = FALSE,
  FPKMensembl = FALSE, outFile)
}
\arguments{
\item{GR}{a genomic ranges object containing gene names in metadata column 'name'}

\item{FPKM}{a data.frame containing FPKM data and gene names of the same format as those in GR$name}

\item{FPKMgene}{the name of the column of FPKM to be compared to GR$name for merging}

\item{columns}{a vector of column names from FPKM with FPKM values to be analyzed}

\item{FPKMsymbol}{the name of the column with gene symbols/names to be added to GR [default:""]}

\item{FPKMensembl}{the name of the column with ensembl gene IDs to be added to GR [default:""]}

\item{outFile}{the name of the file to be created with the results of this function, optional.}
}
\value{
a genomic ranges object with FPKM data
}
\description{
Combine a data.frame with FPKM values and a genomic ranges object
}
\details{
Requires: GenomicRanges

Returns a genomic ranges object with FPKM values (mean of replicates, if given)
Gene symbols or ensemblIDs can also be added to to the object if requested and
and available in the FPKM data.frame
NOTE: current version assumes that length of the genomic ranges object and the 
number of rows in the data.frame are equal and that the gene sets are identical, just
in a different order.
}
\examples{
bed <- rtracklayer::import.bed("geneinfo.bed")
RSEM <- read.table("RSEM.genes.FPKM.all_samples.txt", header=T, stringsAsFactors=F)
out <- add_FPKM_to_GR(GR=bed, FPKM=RSEM, FPKMgene="gene_id", 
   columns=c("S35_FPKM","S36_FPKM"), FPKMsymbol="GeneName",outFile= "geneinfo_FPKM.txt")

}
\author{
Tovah Markowitz \email{tovah.markowitz@nih.gov}
}
\keyword{FPKM}
\keyword{GenomicRanges}
\keyword{RNAseq}
