% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodbnet_annot.R
\name{biodbnet_annot}
\alias{biodbnet_annot}
\title{Annotate genes using bioDBnet}
\usage{
biodbnet_annot(genes = NULL, gfile = NULL, outfile = NULL,
  taxonID = 9606, maxN = 1000, delim = "\\t")
}
\arguments{
\item{gfile}{input text file, one column, each line includes one gene symbol (default=NULL)}

\item{outfile}{optional output text file to write annotation dataframe to (default=NULL)}

\item{taxonID}{taxonomic ID for the gene list (default=9606 for Homo sapiens, use 10090 for Mus musculus)}

\item{maxN}{maximum number of genes to send in each curl command (default=1000)}

\item{delim}{field delimiter used in output text file (default=tab)}

\item{glist}{input character vector of gene symbols (default=NULL)}
}
\value{
annotation dataframe with gene symbols are row names and annotation databases are column names
additionally, will output to file as comma or tab delimited fields
}
\description{
Return annotations from multiple databases included in bioDBnet for a list of genes
}
\details{
Takes a list of gene symbols, constructs a curl command that accesses 
the FNLCR Biological Database Network, bioDBnet, and returns a dataframe
containing annotations from each database for each gene.
Optionally writes out the annotations to a text file.
}
\examples{
biodbnet_annot(gfile="mygenes.txt", outfile="myannots.txt")
myannots <- (genes=mygenes, taxonID=10090)

}
\author{
Justin Lack \email{justin.lack@nih.gov}
}
\keyword{annotation}
\keyword{bioDBnet}
