% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_topTable.R
\name{check_topTable}
\alias{check_topTable}
\title{Check topTable RNASeq results against the expression set for QC}
\usage{
check_topTable(eset, fit, coef, trt, ctrl, n = 10)
}
\arguments{
\item{eset}{an expression set matrix used for creating the fit to be tested}

\item{fit}{the limma fit to be evaluated}

\item{coef}{the fit coefficient (column name from fit$coefficients) to evaluate}

\item{trt}{a vector of column names or indices for subsetting the treatment samples from eset}

\item{ctrl}{a vector of column names or indices for subsetting the control samples from eset}

\item{n}{the number of rows to pull from topTable and report on [default=10]}
}
\value{
a list object of three dataframes: checkTable, trtExpr, ctrlExpr
}
\description{
Compares the differential expression values of two sets of samples (treatment and control) 
as reported in topTable against the underlying expression set to quality control the 
magnitude and direction of the computed fold change.
}
\details{
QC checks: 
    1) does the sample names reported in trtExpr and ctrlExpr match the expected sample names for each set?
    2) does Eset_AveExpr equal Fit_AveExpr exactly?  They should be equal to 6 decimal places
    3) does the direction (+ or -) of Eset_logFC match the direction of Fit_logFC?
    4) does the magnitude of Eset_logFC match the magnitude of Fit_logFC?  
       These will be different but certainly less than 0.5.
    5) Does the variance level in the treatment and in the control samples warrant Fit_pVal reported?

Returns a list object with three dataframes:
    checkTable - the comparison table, with the first n rownames from topTable, 
                 Mean_Trt = average expression level of treatment samples
                 Mean_Ctrl = average expression level of control samples
                 Eset_AveExpr = average expression level of treatment and control samples
                 Fit_AveExpr = average expression level read from the topTable results
                 Eset_logFC = the predicted log2 fold change calculated directly from the expression data
                 Fit_logFC = the calculate log2 fold change for the contrast reported in topTable.
                 Fit_pVal = the p-value reported in topTable
                 
    trtExpr -    a dataframe of the expression levels of each of the treatment samples 
                 for the rows reported in checkTable
                 
    ctrlExpr -   a dataframe of the expression levels of each of the control samples 
                 for the rows reported in check Table
}
\examples{
check_topTable(eset=theEset, fit=theFit, coef="Trt_Ctrl", trt=trtSamples, ctrl=ctrlSamples)
check_topTable(eset=theEset, fit=theFit, coef="KOvsWT",
               trt=theSamples[grepl("KO_", theSamples)], ctrl=theSamples[grepl("WT_", theSamples)])

}
\author{
Susan Huse \email{susan.huse@nih.gov}
}
\keyword{RNASeq}
\keyword{check}
\keyword{limma}
\keyword{topTable}
