% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factordf_to_mtx.R
\name{factordf_to_mtx}
\alias{factordf_to_mtx}
\title{Convert a dataframe with factors to a numeric matrix}
\usage{
factordf_to_mtx(x, cnames = covars, clevels = covar_levels)
}
\arguments{
\item{x}{dataframe containing factor data}

\item{cnames}{character vector of column names to be used as factors}

\item{clevels}{character vector of factor levels, determining the order of the factor levels}
}
\value{
numeric matrix
}
\description{
Reads in a dataframe with along with factor column names and levels
  and returns a numeric matrix populated with the factor level integers
}
\examples{
factordf_to_mtx(groupsDF, cnames=covars, clevels=covar_levels)

}
\author{
Susan Huse \email{susan.huse@nih.gov}
}
\keyword{Pipeliner}
\keyword{RNASeq}
\keyword{limma}
