% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_filtered_topTable.R
\name{get_filtered_topTable}
\alias{get_filtered_topTable}
\title{Filter topTable results by fold change, significance, and n}
\usage{
get_filtered_topTable(theFit, theCoef = NULL, q = 0.05, n = 10,
  lFC = 0.58, addGene = TRUE, miRNA = FALSE)
}
\arguments{
\item{theFit}{a limma fit (MArrayLM) object}

\item{theCoef}{the coefficient in the fit object to analyze (default=NULL)}

\item{q}{minimum FDR / q / adjusted p-value to include (default=0.05)}

\item{n}{maximum number of rows to include in the final output dataframe (default = 10)}

\item{lFC}{minimum log2 fold change to include either up or down (abs(logFC)) (default = 0.58 = FC of 1.5)}

\item{addGene}{parse out the gene name from the gene ID reported with CCBR Pipeliner (default = TRUE)}

\item{miRNA}{boolean, specifying if fit object is an miRNA DGELRT object (default=FALSE)}
}
\value{
dataframe, with the format provided by limma::topTable
}
\description{
Runs limma topTable on an input fit object, 
filters by minimum q-value, logFC, and returns up to n rows.
}
\details{
Runs the limma topTable function, returning all rows, 
then filters out non-hits by minimum q-value (FDR), 
and a minimum, absolute fold change (abs(log2FC)).
The results are ordered by abs(logFC) and then the first "n" rows are returned.

The addGene option will parse out the gene name from the gene ID reported by the CCBR Pipeliner, 
and add the gene name as a new column to the output matrix.

Returns a topTable style dataframe
}
\examples{
myFit <- limma_model(x=eset, y=covarMtx, cols=covars, interact=FALSE)
myHits <- get_filtered_topTable(myFit, theCoef="isKO", q=0.1, n=50, lFC=1)
}
\author{
Susan Huse \email{susan.huse@nih.gov}
}
\keyword{Pipeliner}
\keyword{RNASeq}
\keyword{limma}
