% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_venn_matrix.R
\name{get_venn_matrix}
\alias{get_venn_matrix}
\title{Create a Venn diagram matrix of significant genes}
\usage{
get_venn_matrix(dflist, theNames)
}
\arguments{
\item{dflist}{a list of topTables results, e.g., from get_filtered_topTable}

\item{theNames}{a vector of names for labeling each of the sets.}

\item{q}{minimum FDR / q / adjusted p-value to include (default=0.05)}

\item{n}{maximum number of rows to include in the final output dataframe (default = 10)}

\item{lFC}{minimum log2 fold change to include either up or down (abs(logFC)) (default = 1.5)}

\item{addGene}{parse out the gene name from the gene ID reported with CCBR Pipeliner (default = TRUE)}
}
\value{
a list of two dataframes of the format provided by limma::topTable, the Venn counts and the logFC
}
\description{
Creates matrices of significant differentially expressed genes from a list of topTable results for drawing Venn diagrams
}
\details{
Reads a list of differentially expressed gene comparisons as limma topTable objects,
and creates a unified matrix of differentially expressed genes across the list of contrasts.
The output is designed for use with the limma::vennDiagram function.
NB: the vennDiagram function is limited to 5 sets (contrasts)

The input list of topTable dataframes will be combined into a unified dataframe of fold change values.
NB: It is very important to FILTER FOR SIGNIFICANCE BEFORE running get_venn_matrix.
get_filtered_topTable has been designed to do this prefiltering.

This function will merge the logFC data from each of the input filtered, topTable dataframes
to create a single logFC matrix of all topTable results. 
Genes significant in one contrast but not another will have NA values.

The logFC matrix will be converted to a Venn diagram matrix with values of 1,0,-1, where
significant logFC > 1 = 1, 
significant logFC < 1 = -1, and
not significant logFC = 0.
This Venn matrix can then be used by limma::vennDiagram

Returns a list of two dataframes: [[1]] = Venn matrix, [[2]] = logFC matrix
}
\examples{
Fit1 <- limma_model(x=eset, y=covarMtx, cols=covars, interact=FALSE)
Set1 <- get_filtered_topTable(Fit1, theCoef="isKO", q=0.1, n=50, lFC=1)
Set2 <- get_filtered_topTable(Fit2, theCoef="isKO", q=0.1, n=50, lFC=1)
Set3 <- get_filtered_topTable(Fit3, theCoef="isKO", q=0.1, n=50, lFC=1)
myVenns = get_venn_matrix(list(
               get_filtered_topTable(Fit1, theCoef="isKO_TRUE", q=0.1, n=10000, lFC=1), 
               get_filtered_topTable(Fit2, theCoef="isKO_TRUE", q=0.1, n=10000, lFC=1), 
               get_filtered_topTable(Fit3, theCoef="isKO_TRUE", q=0.1, n=10000, lFC=1)), 
            theNames=c("Set1", "Set2", "Set3"))
vennDiagram(myVenns[[1]], main="Differential Expression across sets", 
            circle.col=c("red", "blue", "green"), include="both")

}
\author{
Susan Huse \email{susan.huse@nih.gov}
}
\keyword{Pipeliner}
\keyword{RNASeq}
\keyword{limma}
