% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head2.R
\name{head2}
\alias{head2}
\title{print the head(obj) in 2 dimensions}
\usage{
head2(obj, x = 6, y = 6)
}
\arguments{
\item{obj}{an R object to print}

\item{x}{the number of items in the first dimension (length or rows) to print}

\item{y}{the number of items in the second dimension (columns or list subitems) to print}
}
\value{
beginning of the input object to be printed or assigned
}
\description{
Prints out the head of an object in multiple dimensions.
The head() function prints out the first 6 rows, but all columns of a data.frame or matrix
head2 will print out a limited number of rows and columns.
}
\details{
If the input object is a matrix or data.frame, it will be truncated in both rows (x) and columns (y).
If the input object is a list, the list object will be limited to the first x items, 
and each item within it will be printed using head2 recursively.
If the input object is a vector, the first x items will be printed.
}
\examples{
allfits <- list()
for(i in names(DE3_list)){allfits[[i]] <- DE3_list[[i]]$fit}
for(i in names(DE4_list)){allfits[[i]] <- DE4_list[[i]]$fit}
allDF <- merge_topTables(allfits, 
                         names=c(names(DE3_list), names(DE4_list)),
                         coefs=c(DE3_coefs, DE4_coefs), 
                         q=0.10, n=NULL, sortby="logFC")

}
\author{
Susan Huse \email{susan.huse@nih.gov}
}
\keyword{head}
