% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_EnsemblGene.R
\name{insert_EnsemblGene}
\alias{insert_EnsemblGene}
\title{Insert columns for EnsemblID and GeneCode}
\usage{
insert_EnsemblGene(df, colname = "rowname", before = TRUE)
}
\arguments{
\item{df}{input dataframe containing a single column of EnsemblID and gene code}

\item{colname}{the column in hte input dataframe containing the EnsemblID and gene code.  Default = "rowname" to use the rownames, rather than a data column.}

\item{before}{insert columns in front of data columns (Default=TRUE)}
}
\value{
dataframe, with the format provided by limma::topTable
}
\description{
Insert separate columns for EnsemblID and gene code into a dataframe with one column containing both IDs
}
\details{
Reads a column of concatenated EnsemblIDs and gene codes, as generated by the CCBR RNASeq Pipeline
and inserts two new columns, one for each identifier.
The colname parameter is used to specify the name of an existing column that contains the concatenated EnsemblID and gene code
Use "rownames" if these are the rownames, rather than a dataframe column.

If before is set to TRUE, then the two new columns will be inserted as the first columns in the output dataframe.
If before is set to FALSE, then the two new columns will be appened at the end, as the last columns in the output dataframe.

Returns a new dataframe with two additional columns, EnsemblID and GeneCode
}
\examples{
myDF <- insert_EnsemblGene(myDF)
myDF <- insert_EnsemblGene(myDF, colname="EnsemblGene", after=TRUE)

}
\author{
Susan Huse \email{susan.huse@nih.gov}
}
\keyword{Ensembl}
\keyword{Pipeliner}
\keyword{RNASeq}
\keyword{code}
\keyword{gene}
\keyword{limma}
