% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_model.R
\name{limma_model}
\alias{limma_model}
\title{Run a limma model on an expressionset and covariates matrix}
\usage{
limma_model(x, y, cols, interact = FALSE)
}
\arguments{
\item{x}{ExpressionSet (can be created with make_filtered_eset)}

\item{y}{covariate matrix (can be created with factordf_to_mtx)}

\item{cols}{character vector of column names to use from the covariate matrix}

\item{interact}{boolean value for whether to include interaction terms between covariates (default=FALSE)}
}
\value{
MArrayLM object ("fit")
}
\description{
Reads an expression set, covariate matrix, covariate names
 and runs the model design (with and without interaction), 
 fits the model, and runs eBayes.  
 Returns the MArrayLM object ("fit")
}
\examples{
dfList <- load_RawCountsGroup(counts_file = file.path(dir_data, file_data), 
                              groups_file = file.path(dir_data, file_groups))
raw <- dfList[[1]]
groupsDF <- dfList[[2]]
rm(dfList)
names(covar_levels) <- covars
covarMtx <- factordf_to_mtx(groupsDF[, covars], cnames=covars, clevels=covar_levels)
eset <- make_filtered_eset(x=raw, y=covarMtx, g=groupsDF$Group, 
                           minc = filter_minCPM, mins = filter_minSampleCount, 
                           nmethod="quantile")
myFit <- limma_model(x=eset, y=covarMtx, cols=covars, interact=FALSE)

}
\author{
Susan Huse \email{susan.huse@nih.gov}
}
\keyword{Pipeliner}
\keyword{RNASeq}
\keyword{limma}
