% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_model_f.R
\name{limma_model_f}
\alias{limma_model_f}
\title{Run a limma model on an expressionset and formula}
\usage{
limma_model_f(x, f)
}
\arguments{
\item{x}{ExpressionSet (can be created with make_filtered_eset)}

\item{f}{formula}
}
\value{
MArrayLM object ("fit")
}
\description{
Reads an expression set and formula
 and runs the model design, fits the model, and runs eBayes.  
 Returns the MArrayLM object ("fit")
}
\examples{
dfList <- load_RawCountsGroup(counts_file = file.path(dir_data, file_data), 
                              groups_file = file.path(dir_data, file_groups))
raw <- dfList[[1]]
groupsDF <- dfList[[2]]
rm(dfList)
names(covar_levels) <- covars
covarMtx <- factordf_to_mtx(groupsDF[, covars], cnames=covars, clevels=covar_levels)
eset <- make_filtered_eset(x=raw, y=covarMtx, g=groupsDF$Group, 
                           minc = filter_minCPM, mins = filter_minSampleCount, 
                           nmethod="quantile")
Genotype <- covarMtx[,"Genotype"]
Treatment <- covarMtx[, "Treatment"]
myF <- as.formula("~ Genotype + Treatment")
myFit <- limma_model(x=eset, f=myF)

}
\author{
Susan Huse \email{susan.huse@nih.gov}
}
\keyword{Pipeliner}
\keyword{RNASeq}
\keyword{limma}
