% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_RawCountsGroup.R
\name{load_RawCountsGroup}
\alias{load_RawCountsGroup}
\title{Creates an ExpressionSet and covariates dataframe from RNASeq pipeline}
\usage{
load_RawCountsGroup(counts_file, groups_file)
}
\arguments{
\item{counts_file}{raw counts file from the RNASeq pipeline (RawCountFile_RSEM_genes.txt)}

\item{metadata_file}{metadata file from RNASeq pipeline, tab-delimited, with header
          Col1 = SampleID, 
          Col2 = Group 
          Col3 = SampleName, 
          Col4 = Genotype, e.g., KO and WT or Tumor and NonTumor (expects 2 values)
          Col5 = Treatment, e.g., Condition1 and Condition2, Treated and Untreated (any number of values)
Note: Group = Treatment + "_" + Genotype}
}
\value{
list object with [[1]] = numeric matrix of raw count data, 
                         [[2]] = dataframe with sample names and covariates
}
\description{
Imports a raw counts file from the RNASeq pipeliner output (RawCountFile_RSEM_genes.txt)
 and a metadata or targets file with covariates, 
 and creates an numeric matrix of raw counts and covariates dataframe for linear modeling with limma.
 Returns a list containing 1) raw counts matrix and 2) metadata/covariates dataframe
 
 See notes on format of metadata file
}
\examples{
dfList <- load_RawCountsGroup(counts_file = file.path(dir_data, file_data), 
                              groups_file = file.path(dir_data, file_groups))
raw <- dfList[[1]]
groupsDF <- dfList[[2]]
rm(dfList)
names(covar_levels) <- covars
covarMtx <- factordf_to_mtx(groupsDF[, covars], cnames=covars, clevels=covar_levels)
eset <- make_filtered_eset(x=raw, y=covarMtx, g=groupsDF$Group, 
                           minc = filter_minCPM, mins = filter_minSampleCount, 
                           nmethod="quantile")
myFit <- limma_model(x=eset, y=covarMtx, cols=covars, interact=FALSE)

}
\author{
Susan Huse \email{susan.huse@nih.gov}
}
\keyword{Pipeliner}
\keyword{RNASeq}
\keyword{limma}
