% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_filtered_eset.R
\name{make_filtered_eset}
\alias{make_filtered_eset}
\title{Creates a filtered expression set from raw counts for limma modeling}
\usage{
make_filtered_eset(x, d, g = NULL, minc = 0.5, mins = 3,
  nmethod = "quantile", plot = TRUE)
}
\arguments{
\item{x}{raw counts matrix (can be created with load_RawCountsGroup)}

\item{d}{design - passed directly to edgeR::filterByExpr (use NULL if using group)}

\item{g}{group - passed directly to edgeR::filterByExpr (default=NULL), if using group set d=NULL}

\item{minc}{integer value for minimum CPM count filtering threshold (default = 0.5)}

\item{mins}{integer value for minimum number of samples passing minc to be included in filtered set (default = 3)}

\item{nmethod}{normalization method to be used by voom (default = "quantile")}
}
\value{
Elist (limma): expression matrix, weights, design, targets
}
\description{
Filters a raw counts matrix using edgeR::filterByExpr, 
 normalizes the filtered data using limma::voom, and
}
\details{
Returns a filtered, normalized, expression set.
}
\examples{
dfList <- load_RawCountsGroup(counts_file = file.path(dir_data, file_data), 
                              groups_file = file.path(dir_data, file_groups))
raw <- dfList[[1]]
groupsDF <- dfList[[2]]
rm(dfList)
names(covar_levels) <- covars
covarMtx <- factordf_to_mtx(groupsDF[, covars], cnames=covars, clevels=covar_levels)
eset <- make_filtered_eset(x=raw, y=covarMtx, g=groupsDF$Group, 
                           minc = filter_minCPM, mins = filter_minSampleCount, 
                           nmethod="quantile")
myFit <- limma_model(x=eset, y=covarMtx, cols=covars, interact=FALSE)

}
\author{
Susan Huse \email{susan.huse@nih.gov}
}
\keyword{Pipeliner}
\keyword{RNASeq}
\keyword{limma}
