% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_bw_across_feature.R
\name{mean_bw_across_feature}
\alias{mean_bw_across_feature}
\title{Calculate the mean bigwig score across a series of features}
\usage{
mean_bw_across_feature(GR, bigwigFile, progressBar = TRUE, outFile)
}
\arguments{
\item{GR}{a genomic ranges object}

\item{bigwigFile}{name/location of the bigwig file}

\item{progressBar}{whether or not to run the progress bar, suggested when running interactive [default: TRUE]}

\item{outFile}{the name of the file to be created with the results of this function, optional.}
}
\value{
a genomic ranges object with new columns: bwMean, bwTotal, nNotZero, score
}
\description{
Requires: rtracklayer, GenomicRanges
}
\details{
This function takes a set of regions(features) given in the form of a genomic ranges
object. For each region, the bigwig values are extracted using rtracklayer. Using this
information, the following data is saved: 1) the mean of the extracted values, 2) the
sum of the extracted values, 3) the number of non-zero values, and 4) the "score" that
is calculated as the sum divided by the non-zero length.
}
\examples{
bed <- rtracklayer::import.bed("geneinfo.bed")
out <- mean_bw_across_feature(GR=bed, bigwigFile="mH2A_D.inputnorm.bigwig", 
    outFile= "mH2A_D_at_genes.txt") 


}
\author{
Tovah Markowitz \email{tovah.markowitz@nih.gov}
}
\keyword{GenomicRanges}
\keyword{bigwig}
\keyword{bw}
\keyword{rtracklayer}
