% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_topTables.R
\name{merge_topTables}
\alias{merge_topTables}
\title{Merge limma:topTable results for multiple comparisons}
\usage{
merge_topTables(fits, names, coefs, q = 0.05, logFC = 1, n = 10,
  sortby = "logFC")
}
\arguments{
\item{fits}{a list of limma fit (MArrayLM) objects}

\item{names}{a vector of names matching the fit objects to be appended with the topTable column names}

\item{coefs}{a vector of coefficient names matching the fit objects to be used by topTable}

\item{q}{maximum FDR / q / adjusted p-value to include (default=0.05)}

\item{logFC}{minimum logFC to include (default=1)}

\item{n}{maximum number of rows to include (default=10), use NULL to include all filtered rows.}

\item{sortby}{ordering of the rows exported. If "logFC" then sort by descreasing abs(logFC), otherwise sort by the specified topTable column.}
}
\value{
dataframe merged from all input topTable(fit, coef) results
}
\description{
Reads a list of limma fit (MArrayLM) objects, runs topTable with the specified parameters
and then merges all of the tables together by geneID
}
\details{
The list of fits, must have a matching list of names for columns (the comparisons), and 
a list of coefficient column names to provide to topTable.
The resulting matrix of genes from all comparisons will then be filtered by adj.P.Val (q), 
sorted, and then the top n rows returned.
The output column names will be designated by the fit name, "_", and the topTable column name, 
e.g., myFitName_logFC, myFitName_P.Value

Filtering can be limited by selecting q=1.0 to accept all genes, irrespective of significance.
The number of rows returned can be set, or to return all rows set n=NULL
Before selecting n rows and before returning the dataframe, the data will be sorted.
Default sorting is abs(logFC) in descending order.  
To sort on a different column, use the standard topTable output column names (e.g., "adj.P.Val" and "P.Value"
}
\examples{
allfits <- list()
for(i in names(DE3_list)){allfits[[i]] <- DE3_list[[i]]$fit}
for(i in names(DE4_list)){allfits[[i]] <- DE4_list[[i]]$fit}
allDF <- merge_topTables(allfits, 
                         names=c(names(DE3_list), names(DE4_list)),
                         coefs=c(DE3_coefs, DE4_coefs), 
                         q=0.10, n=NULL, sortby="logFC")

}
\author{
Susan Huse \email{susan.huse@nih.gov}
}
\keyword{RNASeq}
\keyword{limma}
\keyword{merge}
\keyword{topTable}
