% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_de_volcano.R
\name{plot_de_volcano}
\alias{plot_de_volcano}
\title{Create a volcano plot from a limma fit object}
\usage{
plot_de_volcano(f, coef, title = "Volcano Plot", interactive = TRUE,
  sigVal = "adj.P.Val", maxp = 0.1, minLFC = 1.5,
  colSigHigh = "orangered4", colSigLow = "steelblue",
  colNonLow = "grey50", colNonHigh = "palegreen4", showsub = TRUE,
  miRNA = FALSE)
}
\arguments{
\item{f}{input limma fit object}

\item{coef}{name of the column in theFit$coefficients to plot}

\item{title}{text to use for the plot main title}

\item{interactive}{boolean, if TRUE use plot_ly to create an interactive plot, otherwise use ggplot (default=TRUE)}

\item{sigVal}{column used for determining significance, "adj.P.Val" or "P.Value" (default="adj.P.Val")}

\item{maxp}{maximum FDR or P value of genes considered statistically significant (default=0.1)}

\item{minLFC}{minimum log2 foldchange of genes considered biologically significant (default=1)}

\item{colSigHigh}{color name for the lines and points highlighting significant differential expression (default="red")}

\item{colSigLow}{color name for the points that are significant but have a low FC (default="green")}

\item{colNonLow}{color name for the points that are not significant and have a low FC (default="grey")}

\item{colNonHigh}{color name for the points that are not significant but have a high FC (default="grey")}

\item{showsub}{boolean, if TRUE show subtitle with counts, if FALSE no subtitling}

\item{miRNA}{boolean, specifying if fit object is an miRNA DGELRT object (default=FALSE)}
}
\value{
ggplot object for printing
}
\description{
Creates a volcano plot from a limma fit (MArrayLM) object, using FDR as the y-axis, and coloring significant genes.
}
\details{
The limma fit is plotted using ggplot, 
with a horizontal line for the FDR threshold, 
and two vertical lines for the logFC threshold.  
The plot will have a main title and a subtitle with counts of differentially expressed genes.
}
\examples{
print(plot_de_volcano(f=myFit, coef="TreatmntTRUE", title="Experiment 1"))

}
\author{
Susan Huse \email{susan.huse@nih.gov}
}
\keyword{RNASeq}
\keyword{limma}
\keyword{plot}
\keyword{volcano}
