% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_limma_counts.R
\name{qc_limma_counts}
\alias{qc_limma_counts}
\title{Prints out raw, normalized, and differential expression data comparing two sets of samples for a particular gene}
\usage{
qc_limma_counts(symbolname, A, B, C, theRaw, theEset, theFit, testresults)
}
\arguments{
\item{symbolname}{the gene symbol serving as rownames in the data objects}

\item{A}{a vector of sample names used as column names in the data objects, to compare with B}

\item{B}{a second vector of sample names used as column names in the data objects, to compare with A}

\item{C}{contrast name to be used in calling topTable (coef="C")}

\item{theRaw}{the raw counts data matrix}

\item{theEset}{the eset of filtered and normalized expression data}

\item{theFit}{the limma eBayes fit object}

\item{testResults}{the result of decideTests}
}
\value{
list object with [[1]] = numeric matrix of raw count data, 
                         [[2]] = dataframe with sample names and covariates
}
\description{
For a given gene symbol and two vectors of sample names, 
prints out the raw data, the normalized data, the differential expression data, 
and test results of the model
}
\examples{
eset <- make_filtered_eset(x=raw, y=covarMtx, g=groupsDF$Group, 
                           minc = filter_minCPM, mins = filter_minSampleCount, 
                           nmethod="quantile")
fit <- lmFit(eset, design)
cont.matrix.dTGFb <- makeContrasts(
                     WT_dTGFb = WNP-WTP,
                     KO_dTGFb = KNP-KTP,
                     Diff=(KNP-KTP)-(WNP-WTP), 
                     levels=design
                     )
fit.dTGFb <- contrasts.fit(fit, cont.matrix.dTGFb)
fit.dTGFb <- eBayes(fit.dTGFb)
results.dTGFb <- decideTests(fit.dTGFb, method="nestedF")
qc_counts(symbolname=i, 
          A = rownames(groupsDF[groupsDF$All == "WTP",]), 
          B = rownames(groupsDF[groupsDF$All == "WNP",]), 
          theRaw=raw, theEset=eset, theFit=fit.dTGFb, testresults=dTGFb, isSymbol=TRUE)

}
\author{
Susan Huse \email{susan.huse@nih.gov}
}
\keyword{Pipeliner}
\keyword{RNASeq}
\keyword{limma}
