% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_symbol.R
\name{split_symbol}
\alias{split_symbol}
\title{Splits the combined EnsemblID and gene name}
\usage{
split_symbol(x, s = c("E", "G"))
}
\arguments{
\item{x}{string or character vector of "symbol" from RNASeq pipeline EnsemblID|Gene}

\item{s}{either "E" or "G" to return the cleaned EnsemblID or the Gene name}
}
\value{
string or character vector of the requested names
}
\description{
Splits a character or character vector containing EnsemblID|Gene
 and returns either the EnsemblID (without decimal values) or the gene name
}
\examples{
split_symbol("ENSMUSG00000000001.4|Gnai3", "E")

}
\author{
Susan Huse \email{susan.huse@nih.gov}
}
\keyword{Pipeliner}
\keyword{RNASeq}
\keyword{limma}
