% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topTable2Excel.R
\name{topTable2Excel}
\alias{topTable2Excel}
\title{Export topTable results to Excel or csv/tsv}
\usage{
topTable2Excel(theFit, theCoef, theFile, theSheet = "topTable",
  q = 0.05, sortby = "logFC", append = TRUE, excel = TRUE,
  sep = ",", miRNA = FALSE)
}
\arguments{
\item{theFit}{a limma fit (MArrayLM) object}

\item{theCoef}{the coefficient in the fit object to analyze (default=NULL)}

\item{theFile}{the name of the output file for the exported data}

\item{theSheet}{the name of the sheet (tab) in the output Excel file for the data}

\item{q}{maximum FDR / q / adjusted p-value to include (default=0.05)}

\item{sortby}{ordering of the rows exported.  If "logFC" then sort by descreasing abs(logFC), otherwise sort by FDR.}

\item{append}{should the data be added as a new sheet to an existing file, or should the output file be reinitialized with this as the only sheet (default=TRUE)}

\item{excel}{export to xlsx format, otherwise export to csv format (default = TRUE)}

\item{sep}{if the file format is excel (excel=FALSE), determines what delimiter to use (default=",").  Use "\t" for tab.}

\item{miRNA}{boolean, specifying if fit object is an miRNA DGELRT object (default=FALSE)}
}
\value{
Excel (xlsx or csv) file is saved to disk.
}
\description{
Reads a limma fit (MArrayLM) object, runs topTable with the specified parameters
the exports the dataframe to xlsx, csv, or tab-delimited format.
}
\details{
For the given fit object, as well as the coefficient to analyze, 
topTable2Excel will run topTable returning all results, before filtering by
the maximum FDR (q, adj.P.Val), a minimum logFC, and a maximum number of rows.
If more rows match the FDR and the logFC filtering criteria than the number of rows requested
the data will be sorted by absolute value of logFC in reverse order before selecting the top n lines.

Data can be exported to either .xlsx or to .csv format
Data can be appended to an Excel spreadsheet by creating a new sheet (tab), but data cannot be appended to an existing csv file.

To export all data, set q=1.0

If you run into java out of memory errors:
** Error in .jcall("RJavaTools", "Ljava/lang/Object;", "invokeMethod", cl,  : 
java.lang.OutOfMemoryError: GC overhead limit exceeded **
you can increase the java memory:
options(java.paramters = "-Xmx8000m")
}
\examples{
options(java.parameters = "-Xmx8000m")
top2Excel(theFit=myFit, theCoef=NULL, theFile=Project_results.xlsx, theSheet="MyProject", append=FALSE)

}
\author{
Susan Huse \email{susan.huse@nih.gov}
}
\keyword{RNASeq}
\keyword{csv}
\keyword{excel}
\keyword{limma}
\keyword{topTable}
\keyword{xlsx}
