% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_de_volcano.R
\name{plot_de_volcano}
\alias{plot_de_volcano}
\title{Create a volcano plot from a limma fit object}
\usage{
plot_de_volcano(f, coef, title = "Volcano Plot", maxq = 0.1,
  minLFC = 1, sigcol = "red", showsub = TRUE)
}
\arguments{
\item{f}{input limma fit object}

\item{coef}{name of the column in theFit$coefficients to plot}

\item{title}{text to use for the plot main title}

\item{maxq}{maximum FDR value of genes considered statistically significant (default=0.1)}

\item{minLFC}{minimum log2 foldchange of genes considered biologically significant (default=1)}

\item{sigcol}{color name for the lines and points highlighting significant differential expression (default="red")}

\item{showsub}{boolean, if TRUE show subtitle with counts, if FALSE no subtitling}
}
\value{
ggplot object for printing
}
\description{
Creates a volcano plot from a limma fit (MArrayLM) object, using FDR as the y-axis, and coloring significant genes.
}
\details{
The limma fit is plotted using ggplot, 
with a horizontal line for the FDR threshold, 
and two vertical lines for the logFC threshold.  
The plot will have a main title and a subtitle with counts of differentially expressed genes.
}
\examples{
print(plot_de_volcano(f=myFit, coef="TreatmntTRUE", title="Experiment 1"))

}
\author{
Susan Huse \email{susan.huse@nih.gov}
}
\keyword{RNASeq}
\keyword{limma}
\keyword{plot}
\keyword{volcano}
